config update & sound replace
