+++ addons/orbis_tank_config/scripts/fnc_abramsInit.sqf
@@ -36,7 +36,7 @@ _vehicle setHit ["HitHull", 0.6]; */
 			{
 				if (((_x distance _vehicle) / (speed _x / 3.6) < 0.2) && (_x getRelDir _vehicle < 45 || _x getRelDir _vehicle < 315)) exitWith {
 					// hint "incoming";
-					["RHSUSF_Error", 1, 2] spawn orbis_tank_fnc_repeatSound;
+					["RHSUSF_Error_High", 1, 2] spawn orbis_tank_fnc_repeatSound;
 					sleep 2;
 				};
 			} forEach _nearAmmos;
